+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added `PipelineBuilder`, a more ergonomic way to configure OpenTelemetry with the Application Insights span exporter.
+- Added support semantic conventions for exceptions. Events with the name "exception" are now converted into Exception telemetry.
 
 ## [0.4.0] - 2020-08-14